Latent Time Joint Mixed Effect Models ('ltjmm')

Diseases that progress over long periods of time are often studied by observing cohorts at different stages of disease for shorter periods of time. We apply MCMC sampling to estimate Latent Time Joint Mixed Effect Models (LTJMM) from short-term observations with unknown relative observation times.

References

Gabry J, Češnovar R, Johnson A (2024). _cmdstanr: R Interface to 'CmdStan'_. https://mc-stan.org/cmdstanr/, https://discourse.mc-stan.org.

Landau WM (2024). _instantiate: Pre-Compiled 'CmdStan' Models in R Packages_. R package version 0.2.0, <https://CRAN.R-project.org/package=instantiate>.

Author

Michael Donohue <mdonohue@usc.edu>